Fix error on loading plugin
I could not use this plugin because on nvim startup there was a stack trace. I found out that in the ```get_idestate``` method, lines is expected to be a list of strings, but it was a binary array with ascii values instead. This one-line change does fix the problem for me, but I do not really know python and do not know if this is the right way to do it. But the plugin now works well.